Export different classes and modules. Update tests.
