+++ src/crypto.js
@@ -1,8 +1,58 @@
-import { sha256, ripemd160 } from 'bcrypto'
+import cryptoHash from 'crypto-hashing'
 import base58 from 'bs58'
 import bech32 from 'bech32'
 
-const crypto = {
+/**
+ * Ensure message is in buffer format.
+ * @param {string} message - any string.
+ * @return {string} Returns Buffer of string.
+ */
+function ensureBuffer (message) {
+  if (typeof message === 'string') {
+    message = Buffer.from(message, 'hex')
+  }
+
+  return message
+}
+
+/**
+ * Get hash of a message in hex.
+ * @param {!string} algorithm - Hashing algorithm.
+ * @param {!string|Buffer} message - Message to be hashed.
+ * @return {string} Returns the hash of a string.
+ */
+function hashToHex (algorithm, message) {
+  return cryptoHash(algorithm, ensureBuffer(message)).toString('hex')
+}
+
+/**
+ * Get hash160 of message.
+ * @param {!string|Buffer} message - message in string or Buffer.
+ * @return {string} Returns the hash160 of a string.
+ */
+function hash160 (message) {
+  return hashToHex('hash160', message)
+}
+
+/**
+ * Get sha256 of message.
+ * @param {!string|Buffer} message - message in string or Buffer.
+ * @return {string} Returns the sha256 of a string.
+ */
+function sha256 (message) {
+  return hashToHex('sha256', message)
+}
+
+/**
+ * Get ripemd160 of message.
+ * @param {!string|Buffer} message - message in string or Buffer.
+ * @return {string} Returns the ripemd160 of a string.
+ */
+function ripemd160 (message) {
+  return hashToHex('ripemd160', message)
+}
+
+export {
   /**
    * Base58 object with decode, decodeUnsafe, and encode functions.
    */
@@ -15,42 +65,8 @@ const crypto = {
    */
   bech32,
 
-  /**
-   * Get hash160 of message.
-   * @param {!string} message - any string.
-   * @return {string} Returns the hash160 of a string.
-   */
-  hash160 (message) {
-    const sha256Hashed = this.sha256(message)
-    const ripemd160Hashed = this.ripemd160(sha256Hashed)
-    return ripemd160Hashed
-  },
-
-  /**
-   * Get sha256 of message.
-   * @param {!string} message - any string.
-   * @return {string} Returns the sha256 of a string.
-   */
   sha256,
-
-  /**
-   * Get ripemd160 of message.
-   * @param {!string} message - any string.
-   * @return {string} Returns the ripemd160 of a string.
-   */
   ripemd160,
-
-  /**
-   * Ensure message is in buffer format.
-   * @param {string} message - any string.
-   * @return {string} Returns Buffer of string.
-   */
-  ensureBuffer (message) {
-    if (typeof message === 'string') {
-      message = Buffer.from(message, 'hex')
-    }
-    return message
-  }
+  hash160,
+  ensureBuffer
 }
-
-export default crypto
