+++ src/schema/index.js
@@ -1,0 +1,7 @@
+import Block from './Block.json'
+import Transaction from './Transaction.json'
+
+export {
+  Block,
+  Transaction
+}
